+++ app/containers/ExperiencesPage/index.js
@@ -1,20 +1,20 @@
 import React from 'react';
 import Helmet from 'react-helmet';
 
-import FindoutExperience from './Findout';
-import TDDuoExperience from './3DDuo';
-import LeonPaulExperience from './LeonPaul';
+import Findout from './Findout';
+import Dcdl from './Dcdl';
+import LeonPaul from './LeonPaul';
 
 export default () => (
   <div>
     <Helmet
-      title="Work Experiences"
+      title="Réalisations et expériences professionnelles"
       meta={[
-        { name: 'description', content: 'After 3 years working as a full stack web developper for a data technology / search engine company, I am embracing the freelance experience.' },
+        { name: 'description', content: 'Voir mes travaux et expériences précedentes, réalisés en javascript, en PHP, etc.' },
       ]}
     />
-    <FindoutExperience />
-    <TDDuoExperience />
-    <LeonPaulExperience />
+    <Findout />
+    <Dcdl />
+    <LeonPaul />
   </div>
 );
